Finished textbox, implemented damage type and gold

+++ scripts/items/item_handler.py
@@ -25,6 +25,10 @@ class Item_Handler():
 
 
     def Remove_Item(self, item, delete_item = False):
+        if not item in self.items:
+            print(item)
+            return
+        
         self.items.remove(item)
         self.game.entities_render.Remove_Entity(item)
         if delete_item:
@@ -53,9 +57,12 @@ class Item_Handler():
             if item.Update_Delete_Cooldown():
                 if not item.delete_countdown:
                     self.Remove_Item(item, True)
+                    continue
 
             if item.picked_up:
                 self.items.remove(item)
+                continue
+
 
             if item.sub_type in throwable_weapons:
                 if not item.special_attack:
@@ -63,6 +70,8 @@ class Item_Handler():
                         item.Set_Delete_Countdown(10)
                     continue
                 try:
+                    if not item in self.items:
+                        continue
                     item.Shoot()
                 except Exception as e:
                     print(f"Item is not throwable {e}", item.sub_type)
